Add workaround for fixed gen in scionproto library
